feat(client): move src/features/core to src/core; more consistent with server arch
