+++ app/client/src/core/__mocks__/useIntersectionObserver.ts
