+++ app/client/src/features/accounts/Demos/LoginFormDemo.tsx
@@ -7,7 +7,7 @@ import VisibilityOff from '@mui/icons-material/VisibilityOff';
 import { Form } from '@local/components/Form';
 import { FormContent } from '@local/components/FormContent';
 import { LoadingButton } from '@local/components/LoadingButton';
-import { useForm } from '@local/features/core';
+import { useForm } from '@local/core';
 
 const useStyles = makeStyles((theme) => ({
     link: {
