+++ app/client/src/features/organizations/CreateOrg.tsx
@@ -1,7 +1,7 @@
 import { useMutation, graphql } from 'react-relay';
 
 import type { CreateOrgMutation } from '@local/__generated__/CreateOrgMutation.graphql';
-import { useSnack } from '@local/features/core';
+import { useSnack } from '@local/core';
 import { OrgForm, OrgFormProps } from './OrgForm';
 
 export const CREATE_ORG_MUTATION = graphql`
