+++ app/client/src/features/organizations/OrgForm.tsx
@@ -1,7 +1,7 @@
 import { Button, TextField } from '@mui/material';
 
 import { LoadingButton } from '@local/components/LoadingButton';
-import { useForm } from '@local/features/core';
+import { useForm } from '@local/core';
 import type { CreateOrganization } from '@local/graphql-types';
 import { Form } from '@local/components/Form';
 import { FormContent } from '@local/components/FormContent';
