+++ app/client/src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 import ServerStyleSheets from '@mui/styles/ServerStyleSheets';
-import { themes } from '@local/features/core';
+import { themes } from '@local/core';
 
 export default class MyDocument extends Document {
     render() {
