+++ app/client/src/pages/logout.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router';
 // import { Redirect } from '@local/domains/Logical/Redirect';
 import { logoutMutation } from '@local/__generated__/logoutMutation.graphql';
 import { Loader } from '@local/components/Loader';
-import { useIsClient, useEnvironment } from '@local/features/core';
+import { useIsClient, useEnvironment } from '@local/core';
 import { useUser } from '@local/features/accounts';
 
 const LOGOUT_MUTATION = graphql`
